% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataset.R
\name{generateDataset}
\alias{generateDataset}
\title{Generate a data set that is on the balance - sample size frontier.}
\usage{
generateDataset(frontierObject, N)
}
\arguments{
\item{frontierObject}{An object generated by \code{makeFrontier()}.}

\item{N}{The number of observations left in the exported data set.
If the user selects an undefined point, \code{generateDataset}
returns a dataset from the nearest defined point on the
frontier.}
}
\description{
\code{generateDataset()} allows the user to export a data set
that sits on the frontier.
}
\examples{
data(lalonde)

my.frontier <- matchit(as.formula("treat ~ age + educ + black + hispan"),
                       lalonde,
                       "MahalFrontier")
n <- 300 # Identify the point from which to select the data
matched.data <- generateDataset(my.frontier, N = n)

}
\references{
King, Gary, Christopher Lucas, and Richard
  Nielsen. "The Balance-Sample 
  Size Frontier in Matching Methods for Causal Inference." (2015).
}
