\name{help.matchit}

\alias{help.matchit}

\title{HTML Help for Matchit Commands and Models}

\description{
  The \code{help.matchit} command launches html help for Matchit commands
  and supported models.  The full manual is available online at
  \url{http://gking.harvard.edu/matchit}.
  }

\usage{
help.matchit(object)
}

\arguments{
  \item{object}{a character string representing a Matchit command or model. 
    \code{help.matchit("command")} will take you to an index of Matchit
    commands and \code{help.matchit("model")} will take you to a list of
    models. }
  }

\seealso{The complete document is available online at
  \url{http://gking.harvard.edu/matchit}.  
}

                                                                                                                                                             
\author{
  Daniel Ho <\email{daniel.ho@yale.edu}>;  Kosuke Imai <\email{kimai@princeton.edu}>; Gary King
  <\email{king@harvard.edu}>; Elizabeth Stuart<\email{stuart@stat.harvard.edu}>
}

\keyword{documentation}








