% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{is.matchit}
\alias{is.matchit}
\title{Checks matchit Class}
\usage{
is.matchit(object)
}
\arguments{
\item{object}{any R object}
}
\value{
Returns \code{TRUE} if its argument has class "matchit" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{matchit} object.
}

