% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchIt-package.R
\docType{package}
\name{MatchIt-package}
\alias{MatchIt}
\alias{MatchIt-package}
\title{MatchIt: Nonparametric Preprocessing for Parametric Causal Inference}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Selects matched samples of the original treated and control groups with similar covariate distributions -- can be used to match exactly on covariates, to match on propensity scores, or perform a variety of other matching procedures. The package also implements a series of recommendations offered in Ho, Imai, King, and Stuart (2007) \doi{10.1093/pan/mpl013}. (The 'gurobi' package, which is not on CRAN, is optional and comes with an installation of the Gurobi Optimizer, available at \url{https://www.gurobi.com}.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kosukeimai.github.io/MatchIt/}
  \item \url{https://github.com/kosukeimai/MatchIt}
  \item Report bugs at \url{https://github.com/kosukeimai/MatchIt/issues}
}

}
\author{
\strong{Maintainer}: Noah Greifer \email{noah.greifer@gmail.com} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

Authors:
\itemize{
  \item Daniel Ho \email{daniel.e.ho@gmail.com} (\href{https://orcid.org/0000-0002-2195-5469}{ORCID})
  \item Kosuke Imai \email{imai@harvard.edu} (\href{https://orcid.org/0000-0002-2748-1022}{ORCID})
  \item Gary King \email{king@harvard.edu} (\href{https://orcid.org/0000-0002-5327-7631}{ORCID})
  \item Elizabeth Stuart \email{estuart@jhu.edu} (\href{https://orcid.org/0000-0002-9042-8611}{ORCID})
}

Other contributors:
\itemize{
  \item Alex Whitworth \email{whitworth.alex@gmail.com} [contributor]
}

}
\keyword{internal}
