\name{user.prompt}

\alias{user.prompt}

\title{Pause in demo files}

\description{
  Use \code{user.prompt} while writing demo files to force users to hit
  return before continuing.  
}

\usage{
user.prompt()
}

\seealso{\code{readline}}

\author{Olivia Lau \email{olau@fas.harvard.edu}
}

\examples{
\dontrun{
user.prompt()
}
}

\keyword{file}
